backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]
